Update association_basics.md
Added a warning for behavior when `:dependent` option is used in conjunction with `:through` option. Text is taken from http://apidock.com/rails/ActiveRecord/Associations/ClassMethods/has_many. I was recently confused by this behavior and overlooked the description in the docs. 

It might be nice to include this warning in the "Methods Added by `has_many`" section as well, but this is where I would have looked first.

For additional context, see:

* https://github.com/rails/rails/issues/7618
* http://stackoverflow.com/questions/30629680/rails-isnt-running-destroy-callbacks-for-has-many-through-join-model